+++ src/string/index.js
@@ -3,6 +3,7 @@ import endsWith from './endsWith'
 import { hasUnicode } from './unicode'
 import startsWith from './startsWith'
 import padStart from '../string/padStart'
+import strtr, { replaceByTemplate } from '../string/strtr'
 import camelCase, { pascalCase } from '../string/camelCase'
 import { reTrim } from '../core/vars'
 
@@ -53,4 +54,4 @@ export function clearSpaces(string) {
   return string.toString().replace(/\s+/g, ' ').trim()
 }
 
-export { endsWith, startsWith, padStart, camelCase, pascalCase, hasUnicode }
+export { endsWith, startsWith, padStart, camelCase, pascalCase, hasUnicode, strtr, replaceByTemplate }