+++ src/utils/base64.js
@@ -1,4 +1,5 @@
 import root from '../internal/root'
+import strtr from '../string/strtr'
 
 const utf8ToB64Node = (string) => {
   return Buffer.from(string).toString('base64')
@@ -27,11 +28,34 @@ export function utf8ToB64(string) {
 }
 
 /**
- * Decode string from base-64 to Unicode
+ * Decode from base-64 to Unicode string
  *
  * @param {string} string
  * @returns {string}
  */
 export function b64ToUtf8(string) {
   return decodeURIComponent(escape(b64ToUtf8Fn(string)))
+}
+
+const SYMBOLS_STANDARD = '+/='
+const SYMBOLS_URL_SAFE = '-_~'
+
+/**
+ * Decode from safe-base-64 to Unicode string
+ *
+ * @param {string} string
+ * @return {string}
+ */
+export function b64ToUtf8Safe(string) {
+  return strtr(b64ToUtf8Fn(string), SYMBOLS_STANDARD, SYMBOLS_URL_SAFE)
+}
+
+/**
+ * Encode from Unicode string to safe-base-64
+ *
+ * @param {string} string
+ * @return {string}
+ */
+export function utf8Tob64Safe(string) {
+  return strtr(utf8ToB64Fn(string), SYMBOLS_STANDARD, SYMBOLS_URL_SAFE)
 }