+++ tests/utils/base64.js
@@ -1,4 +1,6 @@
-import { b64ToUtf8, utf8ToB64 } from '../../src/utils'
+import { b64ToUtf8, b64ToUtf8Safe, utf8ToB64, utf8Tob64Safe } from '../../src/utils'
+import forEach from '../../src/core/forEach'
+import flip from '../../src/object/flip'
 
 describe('utf8ToB64', () => {
   it('should return base64-val', () => {
@@ -9,9 +11,33 @@ describe('utf8ToB64', () => {
 })
 
 describe('b64ToUtf8', () => {
-  it('should return  string from base64-val', () => {
+  it('should return string from base64-val', () => {
     expect(b64ToUtf8('MQ==')).toBe('1')
     expect(b64ToUtf8('SGVsbG8gd29ybGQ=')).toBe('Hello world')
     expect(b64ToUtf8('w5DCn8ORwoDDkMK4w5DCssOQwrXDkcKC')).toBe('Привет')
+  })
+})
+
+const list = {
+  1: 'MQ~~',
+  'MQ~~': 'TVF-fg~~',
+  'Привет!': '0J_RgNC40LLQtdGCIQ~~',
+  'Hello world': 'SGVsbG8gd29ybGQ~',
+  '{"message:{"text":"сообщение !"}"}': 'eyJtZXNzYWdlOnsidGV4dCI6ItGB0L7QvtCx0YnQtdC90LjQtSAhIn0ifQ~~',
+}
+
+describe('Utf8Tob64Safe', () => {
+  it('encode original string to safe base64 hash', () => {
+    forEach(list, (hash, text) => {
+      expect(utf8Tob64Safe(String(text))).toBe(hash)
+    })
+  })
+})
+
+describe('b64ToUtf8Safe', () => {
+  it('decode from safe 6ase64 hash to original string', () => {
+    forEach(flip(list), (hash, text) => {
+      expect(b64ToUtf8Safe(String(text))).toBe(hash)
+    })
   })
 })